requirements.txt
fastapi
uvicorn
pillow
google-generativeai